+++ templates/articles.slab.js
@@ -1,28 +1,30 @@
 (function(window) {
 
     const okanjo = window.okanjo;
-    okanjo.ui.engine.registerCss("articles.block2", "@@include(jsStringEscape('articles.block2.css'))", {id: 'okanjo-article-block2'});
+    okanjo.ui.engine.registerCss("articles.slab", "@@include(jsStringEscape('articles.slab.css'))", {id: 'okanjo-article-slab'});
 
-    okanjo.ui.__article_block2 = "@@include(jsStringEscape('articles.block2.mustache'))";
-
-    okanjo.ui.engine.registerTemplate("articles.block2", okanjo.ui.__article_block2, function (model) {
+    // Reuses block2 markup layout, extended css
+    okanjo.ui.engine.registerTemplate("articles.slab", okanjo.ui.__article_block2, function (model) {
         const data = (this._response || { data: { results: [] } }).data || { results: [] };
+        model.blockClasses = ['okanjo-slab'];
         model.articles = data.results;
         model.config = this.config;
         model.instanceId = this.instanceId;
         model.metricChannel = this._metricBase.ch;
         model.metricContext = this._metricBase.cx;
         model.metricParams = okanjo.net.request.stringify(this._metricBase);
+        model.fitImage = 'okanjo-fit';
 
         // Enforce format restrictions
-        this._enforceLayoutOptions();
+        // this._enforceLayoutOptions();
+        this._enforceSlabLayoutOptions();
 
         // Add branding if necessary
         this._registerCustomBranding('.okanjo-article', 'button');
 
         return model;
     }, {
-        css: [/*'okanjo.core',*/ 'articles.block2', 'okanjo.block2' /*, 'okanjo.modal'*/]
+        css: [ 'articles.slab',/* 'articles.block2',*/ 'okanjo.block2' ]
     });
 
 })(window);